+++ andes/core/discrete.py
@@ -1,5 +1,5 @@
 import logging
-from typing import Optional, Union, Tuple
+from typing import Optional, Union, Tuple, List
 from andes.shared import np
 
 logger = logging.getLogger(__name__)
@@ -10,9 +10,10 @@ class Discrete(object):
     Base class for discrete components which exports boolean flags.
     """
 
-    def __init__(self, name=None, tex_name=None):
+    def __init__(self, name=None, tex_name=None, info=None):
         self.name = name
         self.tex_name = tex_name
+        self.info = info
         self.owner = None
         self.export_flags = []
         self.export_flags_tex = []
@@ -99,7 +100,7 @@ class Limiter(Discrete):
     """
     Base limiter class
 
-    This class compares values and sets limit values
+    This class compares values and sets limit values. Exported flags are `zi`, `zl` and `zu`.
 
     Parameters
     ----------
@@ -121,8 +122,8 @@ class Limiter(Discrete):
         Flags for violating the upper limit
     """
 
-    def __init__(self, u, lower, upper, enable=True, name=None, tex_name=None):
-        super().__init__(name=name, tex_name=tex_name)
+    def __init__(self, u, lower, upper, enable=True, name=None, tex_name=None, info=None):
+        super().__init__(name=name, tex_name=tex_name, info=info)
         self.u = u
         self.lower = lower
         self.upper = upper
@@ -258,6 +259,11 @@ class Selector(Discrete):
 
     Names are in `s0`, `s1`, ... and `sn`
 
+    Warnings
+    --------
+    A potential bug when values in different inputs are equal.
+    FIXME: More than one flag will se true in this case.
+
     Examples
     --------
     Example 1: select the largest value between `v0` and `v1` and put it into vmax.
@@ -311,17 +317,50 @@ class Selector(Discrete):
 
 class Switcher(Discrete):
     """
-    Switcher class based on input parameters.
+    Switcher based on an input parameter.
 
     The switch class takes one v-provider, compares the input with each value in the option list, and exports
     one flag array for each option. The flags are 0-indexed.
+
+    Exported flags are named with `_s0`, `_s1`, ..., with a total number of `len(options)`. See the examples
+    section.
+
+    Notes
+    -----
+    Switches needs to be distinguished from Selector.
+
+    Switcher is for generating flags indicating option selection based on an input parameter. Selector is
+    for generating flags at run time based on variable values and a selection function.
+
+    Examples
+    --------
+    The IEEEST model takes an input for selecting the signal. Options are 1 through 6.
+    One can construct ::
+
+        self.IC = NumParam(info='input code 1-6')  # input code
+        self.SW = Switcher(u=self.IC, options=[1, 2, 3, 4, 5, 6])
+
+    If the IC values from the data file ends up being ::
+
+        self.IC.v = np.array([1, 2, 2, 4, 6])
+
+    Then, the exported flag arrays will be ::
+
+        {'IC_s0': np.array([1, 0, 0, 0, 0]),
+         'IC_s1': np.array([0, 1, 1, 0, 0]),
+         'IC_s2': np.array([0, 0, 0, 0, 0]),
+         'IC_s3': np.array([0, 0, 0, 1, 0]),
+         'IC_s4': np.array([0, 0, 0, 0, 0]),
+         'IC_s5': np.array([0, 0, 0, 0, 1])
+        }
     """
+
     def __init__(self, u, options: Union[list, Tuple], name: str = None, tex_name: str = None, cache=True):
         super().__init__(name=name, tex_name=tex_name)
         self.u = u
-        self.option: list = options
+        self.option: Union[List, Tuple] = options
         self.cache: bool = cache
-        self._eval = False  # if the flags has been evaluated
+        self._eval: bool = False  # if the flags has been evaluated
 
         for i in range(len(options)):
             self.__dict__[f's{i}'] = 0
@@ -341,11 +380,11 @@ class Switcher(Discrete):
 
 class DeadBand(Limiter):
     """
-    Deadband with the direction of return.
+    Dead band with the direction of return.
 
     Parameters
     ----------
-    u
+    u : NumParam
         The pre-deadband input variable
     center : NumParam
         Neutral value of the output
@@ -418,6 +457,9 @@ class DeadBand(Limiter):
 
     def check_var(self):
         """
+        Notes
+        -----
+
         Updates five flags: zi, zu, zl; zur, and zlr based on the following rules:
 
         zu:
@@ -425,6 +467,7 @@ class DeadBand(Limiter):
 
         zl:
           1 if u < lower; 0 otherwise.
+
         zi:
           not(zu or zl);