+++ andes/system.py
@@ -293,7 +293,7 @@ class System(object):
             self.vars_to_models()
 
         # store the inverse of time constants
-        self._store_zf()
+        self._store_Tf()
 
     def store_adder_setter(self, models=None):
         """
@@ -702,16 +702,16 @@ class System(object):
         # do nothing for OrderedDict type
         return models
 
-    def _store_zf(self):
+    def _store_Tf(self):
         """
         Store the inverse time constant associated with equations
         """
         for var in self._adders['f']:
             if var.t_const is not None:
-                np.put(self.dae.zf, var.a, var.t_const.v)
+                np.put(self.dae.Tf, var.a, var.t_const.v)
         for var in self._setters['f']:
             if var.t_const is not None:
-                np.put(self.dae.zf, var.a, var.t_const.v)
+                np.put(self.dae.Tf, var.a, var.t_const.v)
 
     def _call_models_method(self, method: str, models: Optional[Union[str, list, Model, OrderedDict]]):
         """