+++ andes/variables/dae.py
@@ -89,9 +89,8 @@ class DAE(object):
         self.m, self.n, self.o = 0, 0, 0
         self.x, self.y, self.z = np.array([]), np.array([]), np.array([])
         self.f, self.g = np.array([]), np.array([])
-        # `self.zf` is the differential equation status array, 0 if the corresponding differential
-        # equation has a time constant of zero. 1 otherwise.
-        self.zf = np.array([])
+        # `self.Tf` is the time-constant array for differential equations
+        self.Tf = np.array([])
 
         self.fx, self.fy = None, None
         self.gx, self.gy = None, None
@@ -154,7 +153,7 @@ class DAE(object):
         self.clear_ijv()
         self.clear_ts()
 
-        self.zf = np.ones(self.n)
+        self.Tf = np.ones(self.n)
 
     def get_size(self, name):
         """
@@ -252,7 +251,7 @@ class DAE(object):
 
         self.f = np.append(self.f, np.zeros(self.n - len(self.f)))
         self.g = np.append(self.g, np.zeros(self.m - len(self.g)))
-        self.zf = np.append(self.zf, np.ones(self.n - len(self.zf)))
+        self.Tf = np.append(self.Tf, np.ones(self.n - len(self.Tf)))
 
     @property
     def xy(self):